# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Gestor de Facturas Finca
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ocr/extract:
    post:
      operationId: extractInvoiceData
      tags: [ocr]
      summary: Extract invoice data from image using AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExtractInvoiceBody"
      responses:
        "200":
          description: Extracted invoice data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExtractedInvoiceData"
  /ocr/validate:
    post:
      operationId: validateInvoiceData
      tags: [ocr]
      summary: Validate and normalize invoice data using AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateInvoiceBody"
      responses:
        "200":
          description: Validated and corrected invoice data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExtractedInvoiceData"
...
  /users/{id}/role:
    patch:
      operationId: updateUserRole
      tags: [users]
      summary: Update a user's role (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRoleBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppUser"
...
components:
  schemas:
...
    ExtractInvoiceBody:
      type: object
      properties:
        imageBase64:
          type: string
      required:
        - imageBase64
    ExtractedInvoiceData:
      type: object
      properties:
        invoiceNumber:
          type: ["string", "null"]
        supplier:
          type: ["string", "null"]
        date:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        totalAmount:
          type: ["number", "null"]
        description:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/ExtractedInvoiceItem"
      required:
        - items
...
    ExtractedInvoiceItem:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        quantity:
          type: ["number", "null"]
        unit:
          type: ["string", "null"]
        unitPrice:
          type: ["number", "null"]
        totalPrice:
          type: ["number", "null"]
      required:
        - name
        - description
    Role:
      type: string
      enum: [viewer, editor, admin]
...
    AppUser:
      type: object
      required: [id, role, isAdminBootstrap]
      properties:
        id: { type: string }
        email: { type: ["string", "null"] }
        name: { type: ["string", "null"] }
        imageUrl: { type: ["string", "null"] }
        role: { $ref: "#/components/schemas/Role" }
        isAdminBootstrap: { type: boolean, description: "True when admin via ADMIN_EMAILS (cannot be downgraded)" }
        createdAt: { type: ["string", "null"] }
        lastSignInAt: { type: ["string", "null"] }
    UpdateUserRoleBody:
      type: object
      required: [role]
      properties:
        role: { $ref: "#/components/schemas/Role" }
...
    ValidateInvoiceBody:
      type: object
      properties:
        invoiceNumber:
          type: ["string", "null"]
        supplier:
          type: ["string", "null"]
        date:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        totalAmount:
          type: ["number", "null"]
        description:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        buyer:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/ExtractedInvoiceItem"
      required:
        - items