# lib/api-zod/src/generated/types/appUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Gestor de Facturas Finca
 * OpenAPI spec version: 0.1.0
 */
import type { Role } from "./role";

export interface AppUser {
  id: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  role: Role;
  /** True when admin via ADMIN_EMAILS (cannot be downgraded) */
  isAdminBootstrap: boolean;
  /** @nullable */
  createdAt?: string | null;
  /** @nullable */
  lastSignInAt?: string | null;
}