# lib/api-zod/src/generated/types/updateUserRoleBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Gestor de Facturas Finca
 * OpenAPI spec version: 0.1.0
 */
import type { Role } from "./role";

export interface UpdateUserRoleBody {
  role: Role;
}